move transaction serializer out of script_score.cpp
